explors how event trigger service might look like

+++ lib/domain/src/main/kotlin/io/hamal/lib/domain/Shard.kt
@@ -20,7 +20,7 @@ data class Shard(val value: Int) {
 
     object Serializer : KSerializer<Shard> {
         override val descriptor: SerialDescriptor
-            get() = PrimitiveSerialDescriptor("Shard", PrimitiveKind.LONG)
+            get() = PrimitiveSerialDescriptor("Shard", PrimitiveKind.INT)
 
         override fun deserialize(decoder: Decoder): Shard {
             return Shard(decoder.decodeInt())
